Consolidating the three minimal_ccl_common files
