+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/kernels/minimal_default_writer.cpp
@@ -11,7 +11,7 @@
 #include "cpp/ttnn/operations/ccl/ccl_host_types.hpp"
 #include "cpp/ttnn/operations/ccl/kernel_common/sharding_addrgen.hpp"
 #include "tt_metal/fabric/hw/inc/tt_fabric_status.h"
-#include "minimal_ccl_common.hpp"
+#include "cpp/ttnn/operations/experimental/ccl/minimal_ccl_common.hpp"
 #include <cstdint>
 #include <utility>
