+++ ttnn/cpp/ttnn/operations/experimental/ccl/reduce_scatter_minimal_async/device/kernels/ring_reduce_scatter_minimal_async_writer.cpp
@@ -13,7 +13,7 @@
 #include "cpp/ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
 #include "tt_metal/fabric/hw/inc/tt_fabric_status.h"
 #include "tt_metal/fabric/hw/inc/linear/addrgen_api.h"
-#include "minimal_ccl_common.hpp"
+#include "cpp/ttnn/operations/experimental/ccl/minimal_ccl_common.hpp"
 #include <cstdint>
 #include <utility>
 
@@ -278,7 +278,8 @@ void kernel_main() {
                                 uint64_t remote_noc0_dest_noc_addr_tile_two =
                                     get_noc_addr(tile_two_id, intermediate_addrgen, 0 /*offset*/, 0 /*noc_id*/);
 
-                                scatter_write_and_advance_local_read_address_for_fabric<
+                                scatter_write_for_fabric_write<
+                                    true,
                                     fabric_mux_num_buffers_per_channel>(
                                     intermediate_addrgen,
                                     tile_one_id,
@@ -302,7 +303,7 @@ void kernel_main() {
                                 uint64_t remote_noc0_dest_noc_addr =
                                     get_noc_addr(tile_id, intermediate_addrgen, 0 /*offset*/, 0 /*noc_id*/);
 
-                                write_and_advance_local_read_address_for_fabric<fabric_mux_num_buffers_per_channel>(
+                                write_for_fabric_write<true,fabric_mux_num_buffers_per_channel>(
                                     intermediate_addrgen,
                                     tile_id,
                                     pkt_hdr,
