+++ ttnn/cpp/ttnn/operations/experimental/ccl/ring_attention_all_gather_async/device/kernels/ring_attention_all_gather_writer.cpp
@@ -8,7 +8,7 @@
 #include "tt_metal/fabric/hw/inc/noc_addr.h"
 #include "cpp/ttnn/operations/ccl/kernel_common/worker_sync_utils.hpp"
 #include "cpp/ttnn/operations/ccl/ccl_host_types.hpp"
-#include "minimal_ccl_common.hpp"
+#include "cpp/ttnn/operations/experimental/ccl/minimal_ccl_common.hpp"
 #include <cstdint>
 #include <utility>
